feat(logout): Added logout on failed refresh token
