+++ src/features/logout/index.ts
@@ -1,1 +1,2 @@
 export * from './ui';
+export * as LogoutModel from './model';
