+++ src/features/logout/ui/LogoutRowButton.tsx
@@ -2,19 +2,13 @@ import { FC } from 'react';
 
 import { useTranslation } from 'react-i18next';
 
-import { IdentityModel } from '@app/entities/identity';
-import { useAppDispatch } from '@app/shared/lib';
-import { SessionModel } from '@entities/session';
 import { RowButton, KeyIcon } from '@shared/ui';
+
+import { useLogout } from '../model';
 
 const LogoutRowButton: FC = () => {
   const { t } = useTranslation();
-  const dispatch = useAppDispatch();
-
-  const logout = () => {
-    dispatch(IdentityModel.actions.logout());
-    SessionModel.clearSession();
-  };
+  const logout = useLogout();
 
   return (
     <RowButton
